Bug 1497903. Update webrender to commit f83c387824b156e7f97f88edee96956bd0de482d
UltraBlame original commit: 13ef28edfdc0ebe2cdddf4947f39a229f6557232

+++ gfx/webrender/src/internal_types.rs
@@ -31,10 +31,12 @@ pub type FastHashSet<K> = HashSet<K, BuildHasherDefault<FxHasher>>;
 
 
 
+
+
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "capture", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
-pub struct CacheTextureId(pub usize);
+pub struct CacheTextureId(pub u64);
